resolve #335
