+++ src/battle/item/UseItem.inc.c
@@ -1,0 +1,107 @@
+#include "common.h"
+
+/// Provide arg `TRUE` on `SI_VAR(1)` to disable refunding.
+EvtSource N(UseItemWithEffect) = SCRIPT({
+    if (SI_VAR(1) == 0) {
+        UseBattleCamPreset(69); // Nice
+        sleep 10;
+
+        PlaySoundAtActor(ACTOR_PLAYER, SOUND_UNKNOWN_208D);
+        SetAnimation(ACTOR_PLAYER, 0, ANIM_GOT_ITEM);
+        GetActorPos(ACTOR_PLAYER, $x, $y, $z);
+        $x += 18;
+        SetActorSpeed(ACTOR_PLAYER, 4.0);
+        SetGoalPos(ACTOR_PLAYER, $x, $y, $z);
+        PlayerRunToGoal(ACTOR_PLAYER);
+
+        $y += 45;
+        $effectY = $y;
+        $effectY += 10;
+        $effectY += 2;
+        PlayEffect(0x33, 1, $x, $effectY, $z, 1.0, 30, 0, 0, 0, 0, 0, 0, 0);
+        MakeItemEntity(SI_VAR(10), $x, $y, $z, 1, 0);
+        SI_VAR(10) = $x;
+
+        N(GiveRefund)();
+        sleep $x;
+
+        sleep 15;
+
+        N(GiveRefundCleanup)();
+        RemoveItemEntity(SI_VAR(10));
+    } else {
+        // No refund.
+
+        GetActorPos(ACTOR_PLAYER, $x, $y, $z);
+        PlaySoundAtActor(ACTOR_PLAYER, SOUND_UNKNOWN_208D);
+        SetAnimation(ACTOR_PLAYER, 0, ANIM_GOT_ITEM);
+        sleep 4;
+
+        $y += 45;
+        $effectY = $y;
+        $effectY += 10;
+        $effectY += 2;
+        PlayEffect(0x33, 1, $x, $effectY, $z, 1.0, 30, 0, 0, 0, 0, 0, 0, 0);
+        MakeItemEntity(SI_VAR(10), $x, $y, $z, 1, 0);
+        SI_VAR(10) = $x;
+
+        sleep 15;
+        RemoveItemEntity(SI_VAR(10));
+    }
+});
+
+EvtSource N(UseItem) = SCRIPT({
+    UseBattleCamPreset(19);
+    SetBattleCamTarget(-85, 1, 0);
+    SetBattleCamOffsetZ(41);
+    SetBattleCamZoom(248);
+    MoveBattleCamOver(30);
+    sleep 10;
+
+    SetAnimation(ACTOR_PLAYER, 0, ANIM_GOT_ITEM);
+    GetActorPos(ACTOR_PLAYER, $x, $y, $z);
+    $y += 45;
+    MakeItemEntity(SI_VAR(10), $x, $y, $z, 1, 0);
+    SI_VAR(14) = $x;
+
+    N(GiveRefund)();
+    sleep $x;
+
+    sleep 15;
+
+    N(GiveRefundCleanup)();
+    RemoveItemEntity(SI_VAR(14));
+});
+
+EvtSource N(PlayerGoHome) = SCRIPT({
+    UseIdleAnimation(ACTOR_PLAYER, 0);
+    SetGoalToHome(ACTOR_PLAYER);
+    SetActorSpeed(ACTOR_PLAYER, 8.0);
+    SetAnimation(ACTOR_PLAYER, 0, ANIM_RUNNING);
+    PlayerRunToGoal(ACTOR_PLAYER);
+
+    SetAnimation(ACTOR_PLAYER, 0, ANIM_10002);
+    UseIdleAnimation(ACTOR_PLAYER, 1);
+});
+
+EvtSource N(EatItem) = SCRIPT({
+    spawn {
+        loop 4 {
+            PlaySoundAtActor(ACTOR_PLAYER, SOUND_UNKNOWN_2095);
+            sleep 10;
+        }
+    }
+    SetAnimation(ACTOR_PLAYER, 0, ANIM_EAT);
+    sleep 45;
+});
+
+EvtSource N(DrinkItem) = SCRIPT({
+    spawn {
+        loop 4 {
+            PlaySoundAtActor(ACTOR_PLAYER, SOUND_UNKNOWN_2095);
+            sleep 10;
+        }
+    }
+    SetAnimation(ACTOR_PLAYER, 0, ANIM_DRINK);
+    sleep 45;
+});
